MOve SJP filtering into its owned class
